Bump version odlparent->6.0.0,mdsal->5.0.3
This adopts new version of MRI projects and adjusts for the changes
therein. Most notably DPs/DPNs are identified by Uint64, not
BigInteger.

+++ cloudscaler/api/src/main/java/org/opendaylight/genius/cloudscaler/api/TombstonedNodeManager.java
@@ -7,11 +7,10 @@
  */
 package org.opendaylight.genius.cloudscaler.api;
 
-import java.math.BigInteger;
 import java.util.List;
 import java.util.function.Function;
-
 import org.opendaylight.controller.md.sal.common.api.data.ReadFailedException;
+import org.opendaylight.yangtools.yang.common.Uint64;
 
 public interface TombstonedNodeManager {
 
@@ -21,13 +20,13 @@ public interface TombstonedNodeManager {
      * @return true if the supllied dpn is getting scaled in
      * @throws ReadFailedException throws read failed exception
      */
-    boolean isDpnTombstoned(BigInteger dpnId) throws ReadFailedException;
+    boolean isDpnTombstoned(Uint64 dpnId) throws ReadFailedException;
 
     /**
      * Add the listener callback which will be invoked upon recovery of scaled in dpn.
      * @param callback callback to be invoked on recovery
      */
-    void addOnRecoveryCallback(Function<BigInteger, Void> callback);
+    void addOnRecoveryCallback(Function<Uint64, Void> callback);
 
     /**
      * Filters the list of dpns which are not scaled in.
@@ -35,5 +34,5 @@ public interface TombstonedNodeManager {
      * @return filtered list of dpns which are not scaled in
      * @throws ReadFailedException throws read failed exception
      */
-    List<BigInteger> filterTombStoned(List<BigInteger> dpns) throws ReadFailedException;
+    List<Uint64> filterTombStoned(List<Uint64> dpns) throws ReadFailedException;
 }
