+++ itm/itm-impl/src/main/java/org/opendaylight/genius/itm/cache/OfEndPointCache.java
@@ -7,35 +7,33 @@
  */
 package org.opendaylight.genius.itm.cache;
 
-import java.math.BigInteger;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import javax.inject.Singleton;
+import org.opendaylight.yangtools.yang.common.Uint64;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @Singleton
 public class OfEndPointCache {
-
     private static final Logger LOG = LoggerFactory.getLogger(OfEndPointCache.class);
 
-    private final ConcurrentMap<BigInteger, String> ofEndPtMap =
-            new ConcurrentHashMap<>();
+    private final ConcurrentMap<Uint64, String> ofEndPtMap = new ConcurrentHashMap<>();
 
-    public void add(BigInteger dpnId, String ofTunnelName) {
+    public void add(Uint64 dpnId, String ofTunnelName) {
         ofEndPtMap.put(dpnId, ofTunnelName);
     }
 
-    public String get(BigInteger dpnId) {
+    public String get(Uint64 dpnId) {
         return ofEndPtMap.get(dpnId);
     }
 
-    public String remove(BigInteger dpnId) {
+    public String remove(Uint64 dpnId) {
         return ofEndPtMap.remove(dpnId);
     }
 
-    public Set<BigInteger> getAll() {
+    public Set<Uint64> getAll() {
         return ofEndPtMap.keySet();
     }
 }
